pyproject.toml
[tool.poetry]
name = "renewal-desk-agent"
version = "0.1.0"
description = "Production-style LLM renewal brief agent"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
psycopg = { version = "^3.1.18", extras = ["binary"] }
opentelemetry-sdk = "^1.25.0"
opentelemetry-exporter-otlp = "^1.25.0"
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true
